requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
anyio==3.7.0
async-timeout==4.0.3
attrs==23.1.0
bidict==0.22.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.3
dnspython==2.3.0
ecdsa==0.18.0
eventlet==0.33.3
exceptiongroup==1.1.3
fastapi==0.95.2
frozenlist==1.4.0
greenlet==2.0.2
gunicorn==21.2.0
h11==0.14.0
httptools==0.6.1
idna==3.4
multidict==6.0.4
Naked==0.1.32
packaging==23.2
passlib==1.7.4
psycopg2==2.9.6
pyasn1==0.5.0
pydantic==1.10.7
pymongo==4.5.0
python-dotenv==1.0.0
python-engineio==4.4.1
python-jose==3.3.0
python-multipart==0.0.6
python-socketio==5.8.0
pywatchman==1.4.1
PyYAML==6.0.1
redis==5.0.0
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.15
starlette==0.27.0
status==0.2.5
transliterate==1.10.2
typing-extensions==4.6.3
urllib3==1.26.6
uvicorn==0.23.2
uvloop==0.19.0
uWSGI==2.0.21
watchfiles==0.21.0
websocket-client==1.6.1
websockets==12.0
yarl==1.9.2
